solve shell
